+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased][unreleased]
+### Bugfix
+* Fixed bugs with parallel mode where providing only one id in an array did not result in correct parallel response return format
 
 ## [3.0.0-rc.10][2016-08-30]
 ### Feature